move reusable snippets to seperate app (to resolve circular dependency issues)
